feat: add helper functions on the project
