+++ src/app/common/helpers/subject.ts
@@ -1,0 +1,9 @@
+export const unicValueSearch: any = (students): any => {
+  const subjects: object = {};
+  students.forEach(item => {
+    Object.keys(item.subjects).forEach(it => {
+      subjects[it] = 1;
+    });
+  });
+  return Object.keys(subjects);
+};
